+++ app/src/androidTest/java/com/pict/pbl/medswift/ExampleInstrumentedTest.kt
@@ -1,12 +1,11 @@
 package com.pict.pbl.medswift
 
-import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
-
+import androidx.test.platform.app.InstrumentationRegistry
+import com.pict.pbl.medswift.symptoms.SymptomsJSONReader
+import org.junit.Assert.*
 import org.junit.Test
 import org.junit.runner.RunWith
-
-import org.junit.Assert.*
 
 /**
  * Instrumented test, which will execute on an Android device.
@@ -15,10 +14,22 @@ import org.junit.Assert.*
  */
 @RunWith(AndroidJUnit4::class)
 class ExampleInstrumentedTest {
+
     @Test
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
         assertEquals("com.pict.pbl.medswift", appContext.packageName)
     }
+
+
+    @Test
+    fun checkSymptomsJsonParsing() {
+        val reader = SymptomsJSONReader( InstrumentationRegistry.getInstrumentation().targetContext )
+        val symptoms = reader.parseSymptoms()
+        for( s in symptoms ) {
+            println( s )
+        }
+    }
+
 }