+++ app/src/main/java/com/pict/pbl/medswift/screens/HomeScreen.kt
@@ -1,4 +1,4 @@
-package com.pict.pbl.medswift
+package com.pict.pbl.medswift.screens
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -8,27 +8,29 @@ import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.tooling.preview.Preview
+import com.pict.pbl.medswift.ui.theme.MedSwiftTheme
 
-
-class MainActivity : ComponentActivity() {
+class HomeScreen : ComponentActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
-            Surface(
-                modifier = Modifier.fillMaxSize(),
-                color = Color.White
-            ) {
-                ActivityUI()
+            MedSwiftTheme() {
+                Surface(
+                    modifier = Modifier.fillMaxSize(),
+                    color = Color.White
+                ) {
+                    ActivityUI()
+                }
             }
         }
     }
 
     @Composable
+    @Preview
     private fun ActivityUI() {
+        // TODO: Add HomeScreen UI here - Bottom Navigation
     }
 
 }
-
-
-