+++ app/src/main/java/com/pict/pbl/medswift/screens/LoginScreen.kt
@@ -1,27 +1,22 @@
 package com.pict.pbl.medswift.screens
 
-import android.graphics.drawable.Icon
+//import androidx.compose.material.icons.fille
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.*
-//import androidx.compose.material.icons.fille
 import androidx.compose.material3.*
 import androidx.compose.runtime.*
-import androidx.compose.runtime.livedata.observeAsState
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.draw.drawBehind
 import androidx.compose.ui.draw.scale
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalConfiguration
@@ -32,8 +27,6 @@ import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.times
-import androidx.constraintlayout.solver.widgets.Rectangle
 import androidx.lifecycle.MutableLiveData
 import com.pict.pbl.medswift.R
 import com.pict.pbl.medswift.login.LoginManager
@@ -58,9 +51,11 @@ class LoginScreen : ComponentActivity() {
             }
 
         }
+        /*
+        TODO: Listen for changes in internet connection, if the internet is off, show a warning to the user
+         */
     }
 
-    //@Preview(showSystemUi = true)
     @Composable
     private fun ActivityUI(){
         val screenConfig = LocalConfiguration.current
@@ -170,10 +165,10 @@ class LoginScreen : ComponentActivity() {
         Button(
             onClick = {
                if( LoginManager.isUserLoggedIn() ) {
-
+                   // TODO: Manage transition to HomeScreen, else show on-boarding screen
                }
                else {
-
+                   // TODO: Manage user login error here
                }
             } ,
             modifier = modifier